refactor : 연동 전으로 변경
