+++ src/main/java/roomescape/domain/Reservation.java
@@ -3,12 +3,13 @@ package roomescape.domain;
 import roomescape.domain.vo.Name;
 
 import java.time.LocalDate;
+import java.time.LocalTime;
 
 public class Reservation {
     private final Long id;
     private final Name name;
     private final LocalDate date;
-    private final ReservationTime time;
+    private final LocalTime time;
 
     public Reservation() {
         this.id = null;
@@ -17,17 +18,27 @@ public class Reservation {
         this.time = null;
     }
 
-    public Reservation(Long id, String name, String date, ReservationTime time) {
+    public Reservation(Long id, String name, String date, String time) {
         this.id = id;
         this.name = new Name(name);
         this.date = translateDate(date);
-        this.time = time;
+        this.time = translateTime(time);
     }
 
     private LocalDate translateDate(String date) {
         // TODO : 검증 로직
 
         return LocalDate.parse(date);
+    }
+
+    private LocalTime translateTime(String time) {
+        // TODO : 검증 로직
+
+        return LocalTime.parse(time);
+    }
+
+    public boolean isEqualId(Long other) {
+        return other == this.id;
     }
 
     public Long getId() {
@@ -42,7 +53,7 @@ public class Reservation {
         return date;
     }
 
-    public ReservationTime getTime() {
+    public LocalTime getTime() {
         return time;
     }
 }
