+++ src/main/java/roomescape/dto/reservation/ReservationResponseDto.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.dto.reservation;
 
 import roomescape.domain.Reservation;
