+++ src/main/resources/schema.sql
@@ -2,9 +2,8 @@ CREATE TABLE reservation (
     id      BIGINT       NOT NULL AUTO_INCREMENT,
     name    VARCHAR(255) NOT NULL,
     date    VARCHAR(255) NOT NULL,
-    time_id BIGINT,
-    PRIMARY KEY (id),
-    FOREIGN KEY (time_id) REFERENCES reservation_time (id)
+    time    VARCHAR(255) NOT NULL,
+    PRIMARY KEY (id)
 );
 
 CREATE TABLE reservation_time (
